Addressed some more of David's code review comments.
